Move configuration-related code to concerns
This makes the field and schema classes less noisy
