+++ spec/unit/app/models/schema_spec.rb
@@ -313,34 +313,6 @@ RSpec.describe Schema do
     end
   end
 
-  describe "#block_display_fields" do
-    describe "when `x-block-display-fields` is set in the schema body" do
-      let(:body) do
-        {
-          "x-block-display-fields" => %w[title description],
-          "properties" => {
-            "title" => {
-              "type" => "string",
-            },
-            "description" => {
-              "type" => "string",
-            },
-          },
-        }
-      end
-
-      it "returns the block display fields from the JSON body" do
-        expect(schema.block_display_fields).to eq(%w[title description])
-      end
-    end
-
-    describe "when `x-block-display-fields` is not set in the schema body" do
-      it "returns an empty array" do
-        expect(schema.block_display_fields).to eq([])
-      end
-    end
-  end
-
   describe "#subschemas_for_group" do
     let(:group_1_subschemas) do
       [
@@ -367,27 +339,6 @@ RSpec.describe Schema do
 
     it "returns an empty array when no subschemas can be found" do
       expect([]).to eq(schema.subschemas_for_group("group_2"))
-    end
-  end
-
-  describe "#embeddable_as_block?" do
-    describe "when `x-embeddable-as-block` is set in the schema" do
-      let(:body) do
-        {
-          "x-embeddable-as-block" => true,
-          "properties" => {},
-        }
-      end
-
-      it "returns true" do
-        expect(schema).to be_embeddable_as_block
-      end
-    end
-
-    describe "when `x-embeddable-as-block` is not set" do
-      it "returns false" do
-        assert_not schema.embeddable_as_block?
-      end
     end
   end
